requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.2.0
langchain-aws>=0.1.0
boto3>=1.34.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-cloudwatch>=1.15.0
opentelemetry-instrumentation-requests>=0.42b0
lxml>=4.9.0
botocore>=1.34.0